Boot Modulation Update
Storage Drivers are no longer in the kernel and are modules that are loaded and released after boot if not used...

+++ include/USRSPC/UserThreads.h
@@ -8,7 +8,7 @@ extern "C"{
 #include <stdatomic.h>
 
 typedef struct {
-	int counter;
+	uint64_t counter;
 }atomic_t;
 
 typedef atomic_t* p_atomic_t;
